Added ashe package and moved Configuration.java to config package, so it is not specifically tied to ashe when move packages are created
