+++ src/main/java/edu/njit/jerse/ashe/api/ApiService.java
@@ -1,4 +1,4 @@
-package edu.njit.jerse.api;
+package edu.njit.jerse.ashe.api;
 
 import java.io.IOException;
 import java.net.http.HttpClient;
