+++ src/main/java/edu/njit/jerse/ashe/models/GPTMessage.java
@@ -1,4 +1,4 @@
-package edu.njit.jerse.models;
+package edu.njit.jerse.ashe.models;
 
 /**
  * Represents a message to or from the ChatGPT API.
